Add files via upload
A Variable in my words is a "word or number" that can be pretty much a hotkey for anything you chose. There are different kinds of variables like int, float, and strings. Int is the most common because it uses less data, it only uses one slot of data and is common to just have a number there. (1) A float is similar to Int but it uses more data and can give you other options to add more code. Like my example earlier, instead of just the number one you can have decimals. (1.001) And a string is where you have a string of different objects all together. Like words, sentences, or a number group. It generally uses 2 or more slots of data. (12345) 
Lets say you are making a game and your main characters name is chosen by the player. You can use these Variables to sync up with the game and allow every time the main characters name appears on the screen a trigger that pulls up the name they chose. That way the player has more flexibility and also the programmer doesnt have to type the name up a million times in the programming process. There's plenty of other uses for Variables but this is the main example I would use it for.